refactor: DiagramEditorに移動。
